refactor: decrease number of files
