+++ src/parsers/css-parser.js
@@ -1,4 +1,7 @@
+const stylis = require('stylis');
 const selectorParser = require('postcss-selector-parser');
+
+const ABELL_CSS_DATA_PREFIX = 'data-abell';
 
 // mostly taken from https://github.com/vuejs/component-compiler-utils/blob/master/lib/stylePlugins/scoped.ts
 // Removed some vue specific features
@@ -36,4 +39,29 @@ const cssSelectorTransformer = (scopingAttribute) => {
 const generateScopedSelector = (selector, attribute) =>
   cssSelectorTransformer(attribute).processSync(selector);
 
-module.exports = generateScopedSelector;
+const selectorPrefixer = (hash) => (element) => {
+  // rule is the AST node which has the css selector
+  if (element.type === 'rule') {
+    // map over each existing select and prefix it
+    const newSelectors = element.props.map((selector) => {
+      return generateScopedSelector(
+        selector,
+        `${ABELL_CSS_DATA_PREFIX}-${hash}`
+      );
+    });
+    // Mutate the old ast node with the new selectors
+    element.props = newSelectors;
+  }
+};
+
+const cssSerializer = (cssString, hash) => {
+  return stylis.serialize(
+    stylis.compile(cssString),
+    stylis.middleware([selectorPrefixer(hash), stylis.stringify])
+  );
+};
+
+module.exports = {
+  cssSerializer,
+  ABELL_CSS_DATA_PREFIX
+};
