+++ src/utils/general-utils.js
@@ -1,5 +1,6 @@
 const fs = require('fs');
 const path = require('path');
+const { ABELL_CSS_DATA_PREFIX } = require('../parsers/css-parser.js');
 
 /**
  * Returns in-built functions from Abell
@@ -95,9 +96,19 @@ const normalizePath = (path) => {
   return path.replace(/\\/g, '/');
 };
 
+// This function uses a single regular expression to add a data prefix to every html opening tag passed to it
+const prefixHtmlTags = (htmlString, hash) => {
+  const openingTagRegexp = /\<([a-zA-Z]+)(.*?)(\s?\/?)\>/gs;
+  return htmlString.replace(
+    openingTagRegexp,
+    `<$1$2 ${ABELL_CSS_DATA_PREFIX}-${hash}$3>`
+  );
+};
+
 module.exports = {
   execRegexOnAll,
   getAbellInBuiltSandbox,
   logWarning,
-  normalizePath
+  normalizePath,
+  prefixHtmlTags
 };
